+++ bin/support/bash_functions.sh
@@ -10,54 +10,6 @@ curl_retry_on_18() {
   return $ec
 }
 
-# This function will install a version of Ruby onto the
-# system for the buildpack to use. It coordinates download
-# and setting appropriate env vars for execution
-#
-# Example:
-#
-#   install_bootstrap_ruby "$BIN_DIR" "$BUILDPACK_DIR"
-#
-# Takes two arguments, the first is the location of the buildpack's
-# `bin` directory. This is where the `download_ruby` script can be
-# found. The second argument is the root directory where Ruby
-# can be installed.
-#
-# This function relies on the env var `$STACK` being set. This
-# is set in codon outside of the buildpack. An example of a stack
-# would be "heroku-24".
-install_bootstrap_ruby()
-{
-  local bin_dir=$1
-  local buildpack_dir=$2
-  local stack="${STACK:?Required env var STACK is not set}"
-
-  # Multi-arch aware stack support
-  if [ "$stack" == "heroku-24" ]; then
-    local arch
-    arch=$(dpkg --print-architecture)
-    local heroku_buildpack_ruby_dir="$buildpack_dir/vendor/ruby/$STACK/$arch"
-  else
-    local heroku_buildpack_ruby_dir="$buildpack_dir/vendor/ruby/$STACK"
-  fi
-
-  # The -d flag checks to see if a file exists and is a directory.
-  # This directory may be non-empty if a previous compile has
-  # already placed a Ruby executable here.
-  if [ ! -d "$heroku_buildpack_ruby_dir" ]; then
-    heroku_buildpack_ruby_dir=$(mktemp -d)
-    # bootstrap ruby
-    "$bin_dir"/support/download_ruby "$bin_dir" "$heroku_buildpack_ruby_dir"
-    function atexit {
-      # shellcheck disable=SC2317
-      rm -rf "$heroku_buildpack_ruby_dir"
-    }
-    trap atexit EXIT
-  fi
-
-  echo "$heroku_buildpack_ruby_dir"
-}
-
 which_java()
 {
   which java > /dev/null